query: add dry run option

+++ query.py
@@ -8,11 +8,11 @@ from db import Search, Paper, Identifier, Keyword, Author
 logger = logging.getLogger(__name__)
 
 # Turn off logging from requests and urllib3 (again)
-#logging.getLogger("requests").setLevel(logging.WARNING)
-#logging.getLogger('urllib3').setLevel(logging.WARNING)
+logging.getLogger("requests").setLevel(logging.WARNING)
+logging.getLogger('urllib3').setLevel(logging.WARNING)
 
 
-def update_search(searchid, session):
+def update_search(searchid, session, dryrun=False):
     search = session.query(Search).filter(Search.id==searchid).one()
     #datefrom:
     if search.last_performed is None:
@@ -22,8 +22,14 @@ def update_search(searchid, session):
         delta = datetime.timedelta(days=31*search.timerange)
         datefrom = search.last_performed - delta
     adsresults = get_ads_results(search.ads_query, datefrom=datefrom, dateto=None)
+    lastperformed = datetime.datetime.now()
     papers = [create_paper_objects(article, search) for article in adsresults]
-    add_papers_to_db(papers, session)
+    add_papers_to_db(papers, session, dryrun=dryrun)
+
+    # Update search with last performed date (do this last in case of issues)
+    if not dryrun:
+        search.last_performed = lastperformed
+        session.commit()
 
 # Does not have a test???
 def get_ads_results(querystring, datefrom=None, dateto=None):
@@ -106,13 +112,15 @@ def create_paper_objects(article, search):
                   searches=[search])
     return paper
 
-def add_papers_to_db(papers, session):
+def add_papers_to_db(papers, session, dryrun=False):
     """
     Add a list of paper objects to database.
 
     Check if they are new, add them if so. If they already exist,
     check if any information has changed and update with new
     information as necessary.
+
+    Update the search
 
     """
 
@@ -124,18 +132,26 @@ def add_papers_to_db(papers, session):
 
         # If no match, add paper as new paper
         if not match:
-            session.add(paper)
-            logger.debug('Added {}: to database'.format(paper.title))
+            if not dryrun:
+                session.add(paper)
+                logger.debug('Added {}: to database'.format(paper.bibcode))
+            else:
+                logger.debug('DRYRUN: would have added {} to database'.format(paper.bibcode))
             new.add(paper.bibcode)
+
         # If there is a match, check if it needs to be updated.
         else:
-            updated.update(check_and_update(match,paper, session))
-    session.commit()
-    logger.info('Added {} new papers to database.'.format(len(new)))
-    logger.info('Update {} existing papers in database.'.format(len(updated)))
+            updated.update(check_and_update(match,paper, session, dryrun=dryrun))
+    if not dryrun:
+        session.commit()
+        logger.info('Added {} new papers to database.'.format(len(new)))
+        logger.info('Update {} existing papers in database.'.format(len(updated)))
+    if dryrun:
+        logger.info('DRYRUN: would have added {} new papers to database.'.format(len(new)))
+        logger.info('DRYRUN: would have Updated {} existing papers in database.'.format(len(updated)))
 
 
-def check_and_update(match, paper, session):
+def check_and_update(match, paper, session, dryrun=False):
     """
     Check if a new version of a paper needs to overwrite an old one.
 
@@ -146,48 +162,59 @@ def check_and_update(match, paper, session):
     # Check single values:
     # bibcode, title, abstract, doi, pub_opeanccess, refered.
     updateset=set()
-    print('Updating title, bibcode, abstract, pub etc for {}'.format(match.title))
+
+    logger.debug('Updating title, bibcode, abstract, pub etc for {}: {}'.format(match.bibcode, match.title))
     for attrib in ['title', 'bibcode', 'abstract', 'pub_openaccess',
                    'refereed', 'doi', 'pubdate']:
         if getattr(match, attrib) != getattr(paper, attrib):
-            setattr(match, attrib, getattr(paper, attrib))
+            if not dryrun:
+                setattr(match, attrib, getattr(paper, attrib))
             updateset.add(attrib)
 
     if (set([i.identifier for i in paper.identifiers]) !=
         set([i.identifier for i in match.identifiers])):
-        for i in match.identifiers:
-            session.delete(i)
-        match.identifiers = paper.identifiers.copy()
+        if not dryrun:
+            for i in match.identifiers:
+                session.delete(i)
+            match.identifiers = paper.identifiers.copy()
         updateset.add('identifiers')
 
     if (set([i.keyword for i in paper.keywords]) !=
         set([i.keyword for i in match.keywords])):
-        for i in match.keywords:
-            session.delete(i)
-        match.keywords = paper.keywords.copy()
+        if not dryrun:
+            for i in match.keywords:
+                session.delete(i)
+            match.keywords = paper.keywords.copy()
+
         updateset.add('keywords')
 
     if (set([i.property for i in paper.properties]) !=
         set([i.property for i in match.properties])):
-        for i in match.properties:
-            session.delete(i)
-        match.properties = paper.properties.copy()
+        if not dryrun:
+            for i in match.properties:
+                session.delete(i)
+            match.properties = paper.properties.copy()
+
         updateset.add('properties')
     if (set([(i.position_,i.author) for i in paper.authors]) !=
         set([(i.position_,i.author) for i in match.authors])):
-        for i in match.authors:
-            session.delete(i)
-        match.authors = paper.authors.copy()
+        if not dryrun:
+            for i in match.authors:
+                session.delete(i)
+            match.authors = paper.authors.copy()
         updateset.add('authors')
 
     # Searches: this should include all, not replace
     if (set([i.id for i in paper.searches]) !=
         set([i.id for i in match.searches])):
-        match.searches += paper.searches
+        if not dryrun:
+            match.searches += paper.searches
+    if not dryrun:
+        session.flush()
+        session.commit()
+        session.flush()
 
-    session.flush()
-    session.commit()
-    session.flush()
+
     return updateset
     ""
 
@@ -208,7 +235,7 @@ def check_if_paper_in_db(paper, session):
     # Add check for multiple results? Don't know if this ever can
     # happen?
     if match:
-        logger.debug('{} already in DB'.format(paper.title))
+        logger.debug('{}: {} already in DB'.format(paper.bibcode,paper.title))
     return match
